formatting and sendout_email cleanups

+++ bbs/instmsg.cpp
@@ -82,7 +82,7 @@ void send_inst_str1(int m, int whichinst, const char *pszSendString) {
   ih.from_user = static_cast<unsigned short>(session()->usernum);
   ih.msg_size = strlen(szTempSendString) + 1;
   ih.dest_inst = static_cast<unsigned short>(whichinst);
-  ih.daten = static_cast<unsigned long>(time(nullptr));
+  ih.daten = static_cast<uint32_t>(time(nullptr));
 
   send_inst_msg(&ih, szTempSendString);
 }
@@ -104,7 +104,7 @@ void send_inst_shutdown(int whichinst) {
   ih.from_user = static_cast<unsigned short>(session()->usernum);
   ih.msg_size = 0;
   ih.dest_inst = static_cast<unsigned short>(whichinst);
-  ih.daten = static_cast<unsigned long>(time(nullptr));
+  ih.daten = static_cast<uint32_t>(time(nullptr));
 
   send_inst_msg(&ih, nullptr);
 }
@@ -125,7 +125,7 @@ void send_inst_cleannet() {
     ih.from_user = 1;
     ih.msg_size = 0;
     ih.dest_inst = 1;
-    ih.daten = static_cast<unsigned long>(time(nullptr));
+    ih.daten = static_cast<uint32_t>(time(nullptr));
 
     send_inst_msg(&ih, nullptr);
   }
@@ -487,7 +487,7 @@ void write_inst(int loc, int subloc, int flags) {
     } else {
       ti.user = 1;
     }
-    ti.inst_started = static_cast<unsigned long>(time(nullptr));
+    ti.inst_started = static_cast<uint32_t>(time(nullptr));
     re_write = true;
   }
 
@@ -578,7 +578,7 @@ void write_inst(int loc, int subloc, int flags) {
     re_write = true;
   }
   if (re_write && syscfg.datadir != nullptr) {
-    ti.last_update = static_cast<unsigned long>(time(nullptr));
+    ti.last_update = static_cast<uint32_t>(time(nullptr));
     File instFile(syscfg.datadir, INSTANCE_DAT);
     if (instFile.Open(File::modeReadWrite | File::modeBinary | File::modeCreateFile)) {
       instFile.Seek(static_cast<long>(session()->GetInstanceNumber() * sizeof(instancerec)), File::seekBegin);
