+++ bbs/shortmsg.cpp
@@ -135,7 +135,7 @@ void SendRemoteShortMessage(int nUserNum, int nSystemNum, char *pszMessageText) 
   nh.main_type = main_type_ssm;
   nh.minor_type = 0;
   nh.list_len = 0;
-  nh.daten = static_cast<unsigned long>(time(nullptr));
+  nh.daten = static_cast<uint32_t>(time(nullptr));
   if (strlen(pszMessageText) > 80) {
     pszMessageText[80] = '\0';
   }
